Fixed bug, new O(1) space solution
Original was giving wrong output for test case. There are actually 5 islands, not 6.  The problem was that it was actually not doing a DFS, instead only looking at the neighbors.

Also implemented a new solution that got away with the visited set. Instead, it assigns "count" to the current value as it is traversing the island.